security in fileuploader.

+++ core/src/main/java/com/thundashop/core/filemanager/FileManager.java
@@ -6,6 +6,7 @@ import com.thundashop.core.common.ManagerBase;
 import com.thundashop.core.databasemanager.data.DataRetreived;
 import com.thundashop.core.usermanager.UserManager;
 import com.thundashop.core.usermanager.data.User;
+import com.thundashop.core.uuidsecuritymanager.UUIDSecurityManager;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
@@ -26,6 +27,9 @@ public class FileManager extends ManagerBase implements IFileManager {
     
     @Autowired
     UserManager UserManager;
+    
+    @Autowired
+    UUIDSecurityManager uUIDSecurityManager;
 
     @Override
     public void dataFromDatabase(DataRetreived data) {
@@ -40,7 +44,11 @@ public class FileManager extends ManagerBase implements IFileManager {
     @Override
     public List<FileEntry> getFiles(String listId) {
         List<FileEntry> res = new ArrayList();
-        res = entries.values().stream().filter(o -> o.listId.equals(listId)).collect(Collectors.toList());
+        for(FileEntry entry : entries.values()) {
+            if(uUIDSecurityManager.hasAccess(entry.id, true, false)) {
+                res.add(entry);
+            }
+        }
         
         Collections.sort(res, new Comparator<FileEntry>(){
             public int compare(FileEntry o1, FileEntry o2){
@@ -54,6 +62,11 @@ public class FileManager extends ManagerBase implements IFileManager {
 
     @Override
     public String addFileEntry(String listId, FileEntry entry) {
+        if(entry.id != null || !entry.id.isEmpty()) {
+            if(!uUIDSecurityManager.hasAccess(entry.id, true, false)) {
+                return null;
+            }
+        }
         entry.listId = listId;
         entry.storeId = storeId;
         saveObject(entry);
@@ -68,6 +81,9 @@ public class FileManager extends ManagerBase implements IFileManager {
 
     @Override
     public void deleteFileEntry(String fileId) {
+        if(!uUIDSecurityManager.hasAccess(fileId, false, true)) {
+            return;
+        }
         FileEntry toRemove = entries.get(fileId);
         if(toRemove != null) {
             deleteObject(toRemove);
@@ -77,6 +93,9 @@ public class FileManager extends ManagerBase implements IFileManager {
 
     @Override
     public FileEntry getFile(String fileId) {
+        if(!uUIDSecurityManager.hasAccess(fileId, true, false)) {
+            return null;
+        }
         return entries.get(fileId);
     }
